# apps/studio/openapi.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "Studio API",
    "version": "1.0"
  },
  "paths": {
    "/api/forge/plan": {
      "post": {
        "summary": "Generate a plan (steps) for a goal using LLM structured output",
        "tags": [
          "ai"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "goal": {
                    "type": "string"
                  },
                  "graphSummary": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Plan with steps array"
          },
          "400": {
            "description": "Invalid request (e.g. goal required)"
          },
          "503": {
            "description": "OpenRouter not configured"
          }
        }
      }
    },
    "/api/graphs/{id}": {
      "get": {
        "summary": "Get graph by ID",
        "tags": [
          "graphs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Graph document"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "patch": {
        "summary": "Update graph",
        "tags": [
          "graphs"
        ],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "flow": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Updated graph document"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/graphs": {
      "get": {
        "summary": "List graphs",
        "tags": [
          "graphs"
        ],
        "responses": {
          "200": {
            "description": "List of forge graph documents"
          },
          "500": {
            "description": "Server error"
          }
        }
      },
      "post": {
        "summary": "Create graph",
        "tags": [
          "graphs"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "title": {
                    "type": "string"
                  },
                  "flow": {}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Created graph document"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
    "/api/image-generate": {
      "post": {
        "summary": "Generate image from prompt via OpenRouter (image modality)",
        "tags": [
          "ai"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "prompt": {
                    "type": "string"
                  },
                  "aspectRatio": {
                    "type": "string"
                  },
                  "imageSize": {
                    "type": "string"
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Base64 data URL of generated image"
          },
          "400": {
            "description": "prompt required"
          },
          "503": {
            "description": "OpenRouter not configured"
          }
        }
      }
    },
    "/api/me": {
      "get": {
        "summary": "Get current user (auth)",
        "tags": [
          "auth"
        ],
        "responses": {
          "200": {
            "description": "User object or null"
          },
          "401": {
            "description": "Not authenticated (body has user null)"
          },
          "500": {
            "description": "Server error"
          }
        }
      }
    },
...
  },
...
}